+++ src/redapid/object.h
@@ -1,6 +1,6 @@
 /*
  * redapid
- * Copyright (C) 2014 Matthias Bolte <matthias@tinkerforge.com>
+ * Copyright (C) 2014-2015 Matthias Bolte <matthias@tinkerforge.com>
  *
  * object.h: Object implementation
  *
@@ -74,7 +74,7 @@ const char *object_get_type_name(ObjectType type);
 bool object_is_valid_type(ObjectType type);
 
 APIE object_create(Object *object, ObjectType type, Session *session,
-                   uint16_t create_flags, ObjectDestroyFunction destroy,
+                   uint32_t create_flags, ObjectDestroyFunction destroy,
                    ObjectSignatureFunction signature);
 void object_destroy(Object *object);
